skip_remote_state_provision option on destroy

+++ qhub/destroy.py
@@ -7,7 +7,7 @@ from .state import terraform_state_sync
 logger = logging.getLogger(__name__)
 
 
-def destroy_configuration(config):
+def destroy_configuration(config, skip_remote_state_provision=False):
     logger.info(
         """Removing all infrastructure, your local files will still remain, \n
     you can use 'qhub deploy' to re - install infrastructure using same config file"""
@@ -24,8 +24,10 @@ def destroy_configuration(config):
         # 03 Remove terraform backend remote state bucket
         # backwards compatible with `qhub-config.yaml` which
         # don't have `terraform_state` key
-        if (config.get("terraform_state", {}).get("type") == "remote") and (
-            config.get("provider") != "local"
+        if (
+            (not skip_remote_state_provision)
+            and (config.get("terraform_state", {}).get("type") == "remote")
+            and (config.get("provider") != "local")
         ):
             terraform_state_sync(config)
             terraform.destroy(directory="terraform-state")
